Revert "CSS: Make show/hide/toggle methods a module"
This reverts commit 67d7a2eefee768b59eb3d51cb1fb2c671873e58a.
