+++ test/unit/basic.js
@@ -54,19 +54,11 @@ QUnit.test( "attributes", function( assert ) {
 
 if ( jQuery.css ) {
 QUnit.test( "css", function( assert ) {
-	assert.expect( 1 );
+	assert.expect( 3 );
 
 	var div = jQuery( "<div/>" ).appendTo( "#qunit-fixture" );
 
 	assert.strictEqual( div.css( "width", "50px" ).css( "width" ), "50px", ".css getter/setter" );
-} );
-}
-
-if ( jQuery.fn.show && jQuery.fn.hide ) {
-QUnit.test( "show/hide", function( assert ) {
-	assert.expect( 2 );
-
-	var div = jQuery( "<div/>" ).appendTo( "#qunit-fixture" );
 
 	div.hide();
 	assert.strictEqual( div.css( "display" ), "none", "div hidden" );
